refactor: clean up and reorganize top-level CLI files
- Remove unused github_integration.py (152 lines)
- Move cli_visualization.py to cli_modules/utils/visualization.py (363 lines)
- Move cli_config.py functions to cli_modules/utils/config_utils.py (295 lines)
- Move cli_auth.py functions to cli_modules/utils/auth_utils.py (262 lines)
- Update all imports to use new module locations
- Remove original cli_config.py and cli_auth.py files
- Fix test imports for relocated functions

Total cleanup: 1,072 lines reorganized into proper module structure
